pyproject.toml
[project]
name = "openai-agent-sdk-with-ollama-locally"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.13.4",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]
